+++ frontend/constants/charts.js
@@ -1,0 +1,10 @@
+export default {
+  oneMonth: '30D',
+  threeMonth: '3M',
+  oneYear: '1Y',
+  max: 'Max',
+  oneMonthQueryLimit: 30,
+  threeMonthQueryLimit: 90,
+  oneYearQueryLimit: 12,
+  maxQueryLimit: 12,
+}